Moved each tab into its on file
